Shift code into app/src Add autoloading bootstrapping
