+++ app/src/Page.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace {
+
+    use SilverStripe\CMS\Model\SiteTree;
+
+    class Page extends SiteTree
+    {
+        private static $db = [];
+
+        private static $has_one = [];
+    }
+}
