--- mysite/code/Page.php
@@ -1,10 +1,0 @@
-<?php
-
-use SilverStripe\CMS\Model\SiteTree;
-
-class Page extends SiteTree
-{
-    private static $db = [];
-
-    private static $has_one = [];
-}
